Now generating SpatialDiff for CBLCM Integration. Properties file, if Cblcm is being used, need to include the following two new poroperties:
cblcm.base.year                    = 2010
cblcm.base.file                      = input/spatialResult2010.csv

Also, the input file spatialResult2010 or any other file containing the baseYear must be provided. The script searches for the base and final year spatial data in both the outputSpatialFile and the cblcm.base.file for both the years in the increasing order of precedence.

Closes #9