Replace test comparison file to fix test; small changes
The new plans output (a) does not have any longer the expected arrival
time (good).  It (b) also has slightly different travel distance
predictions.  I didn't trace it back, but we indeed at some point
included initial and/or final links so it is no wonder that these
distances are a bit longer now.  They are not used by the mobsim
anyways.