I think I found the problem.
Because of some changes, the test results were written to a different location.  Locally, however, results from previous tests at the same location had survived, so the test passed.  On the container infrastructure of travis, obviously that did not work.  I cannot say why the output at that place is so cryptic (IOException) instead of file not found ... I guess it has to do with the fact that the CRCChecksum reader tries to retrieve from the class loader if it does not find it on the file system, and then dies rather silently.  :-(